# backend/src/apps/idc_research/openapi-idc-research.yaml
openapi: 3.0.3
info:
  title: IDC Research APIs
  version: 1.0.0
  description: |
    API specification for IDC Research module.
    Base prefix: `/api`.
    Mounted routers:
      - `/v1/idc-research` (analysis, mock data)
      - `/v1/idc-reports` (reports management, mock data)
servers:
  - url: /api
...
paths:
...
  /v1/idc-research/hardware/products:
    get:
      tags: [idc-research]
      summary: List hardware products
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Category filter (e.g. server/network/storage/os/database/middleware/security)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse_HardwareProducts'
...
  /v1/idc-research/server-failures:
    get:
      tags: [idc-research]
      summary: Server failures aggregated by month/IDC/brand
      parameters:
        - in: query
          name: months
          schema:
            type: integer
            minimum: 1
            maximum: 24
            default: 12
          description: Number of months to include
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse_ServerFailures'
...
components:
  schemas:
...
    UnifiedResponse_HardwareProducts:
      type: object
      properties:
        status: { type: string, example: ok }
        code: { type: integer, example: 200 }
        msg: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/HardwareProduct' }
...
    UnifiedResponse_ServerFailures:
      type: object
      properties:
        status: { type: string, example: ok }
        code: { type: integer, example: 200 }
        msg: { type: string }
        data:
          type: object
          properties:
            monthlyTroubles:
              type: array
              items:
                type: object
                additionalProperties: true
            idcMapping:
              type: object
              additionalProperties: { type: string }
            brandCategories:
              type: object
              additionalProperties: { type: string, enum: [domestic, foreign] }
...
    HardwareProduct:
      type: object
      properties:
        id: { type: string }
        category: { type: string }
        name: { type: string }
        brand: { type: string }
        isDomestic: { type: boolean }
        model: { type: string }
        quantity: { type: integer }
        idcId: { type: string }
        installDate: { type: string, format: date }
        warrantyEndDate: { type: string, format: date }
        status: { type: string }
        failureCount: { type: integer }
        mtbf: { type: integer }
...